Add ConnectWorkerAsync() factory method
This exposes the way a consumer probably wants to use Beanstalk: Reserve
jobs in a loop, one at a time, process them, then reserve the next job.

The IConsumer interface is not exposed to a worker's lambda, so it can
release, touch, and bury jobs, but not reserve another one (which will
interfere with a delete).