Remove all methods from IWorker except Delete, Release, Bury, and Touch
This in preparation for allowing multiple workers to be spawned
on a single connection. Workers shouldn't have logic based on the results
of something like JobStatisticsAsync because if another worker is waiting
for a "reserve" command to return, the result of "stats-job" won't come
back until the job is almost ready to time out.