Deserialization Exception improvements
When a ReserveAsync<T> deserialization failed, the job was reserved, but
the caller did not receive the Job Id, so the caller could not resolve
the error. Now a custom DeserializationException is thrown that has a
Job property. The job is left reserved, but the caller can catch the
exception and use its Job property to resolve the error as they desire:
log the error, send notifications, release/bury/delete the job,
deserialize the data to a different object, etc. Also, to improve code
reuse, a few primary private overloads were added for the other
overloads to call.